Fix tagging and download
The output template format for youtube-dl means that previously we were
downloading the webm file and naming it with a .mp3 extension. When
youtube-dl attempted to use ffmpeg to convert the webm to an mp3 it
failed as the file already existed.

I learned mopidy can play webm, but does not read id3 tags from webm
files, evidently.